+++ CHANGELOG.rst
@@ -1,9 +1,13 @@
 Changelog
 =========
 
+0.0.3
+-----
+#. Included 1875 - `Auto-create Django Admin User During syncdb <http://djangosnippets.org/snippets/1875/>`_
+
 0.0.2
 -----
-#. Included 1031 - http://djangosnippets.org/snippets/1031/
+#. Included 1031 - `Model Inheritance With Content Type <http://djangosnippets.org/snippets/1031/>`_
 
 0.0.1
 -----